Define Node.__iter__()
